Refactor Convert into its own Component
